move internal/search/query to internal/query
