+++ cmd/frontend/graphqlbackend/search_results_test.go
@@ -24,8 +24,8 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/endpoint"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	searchbackend "github.com/sourcegraph/sourcegraph/internal/search/backend"
-	"github.com/sourcegraph/sourcegraph/internal/search/query"
-	searchquerytypes "github.com/sourcegraph/sourcegraph/internal/search/query/types"
+	"github.com/sourcegraph/sourcegraph/internal/query"
+	searchquerytypes "github.com/sourcegraph/sourcegraph/internal/query/types"
 	"github.com/sourcegraph/sourcegraph/internal/search/search"
 )
