+++ internal/query/pattern_type.go
