+++ internal/query/pattern_type_test.go
