+++ internal/query/searchquery.go
@@ -5,8 +5,8 @@ package query
 import (
 	"strings"
 
-	"github.com/sourcegraph/sourcegraph/internal/search/query/syntax"
-	"github.com/sourcegraph/sourcegraph/internal/search/query/types"
+	"github.com/sourcegraph/sourcegraph/internal/query/syntax"
+	"github.com/sourcegraph/sourcegraph/internal/query/types"
 )
 
 // All field names.
