+++ internal/query/syntax/NOTICE
