+++ internal/query/syntax/doc.go
