+++ internal/query/syntax/scanner.go
