+++ internal/query/syntax/scanner_test.go
