+++ internal/query/types/query.go
@@ -6,7 +6,7 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/sourcegraph/sourcegraph/internal/search/query/syntax"
+	"github.com/sourcegraph/sourcegraph/internal/query/syntax"
 )
 
 // A Query is the typechecked representation of a search query.
